Refactored useHistory into separate file
